Remove Unused Validation Checks and HTTP Status Check
This removes several stub validation methods that only returned OK.
This also removes the checkUrl parameters from the Config UI.

Checks to confirm a given URL returns a 2xx response were removed. This
causes confusion when the URL returns a redirect that is supported, but
not properly handled by this very basic check.