Change Retry, Handle 409 Code, and Max Retry is Failure
The retry loop now sleeps for 5 seconds instead of 2. This allows 15
seconds for a previous deployment to finish.

The HTTP code of 409 is used by Marathon to signal when an application
already has a deployment in progress. Prior to this commit, it was
treated as a failure and did not trigger retry.

If the max retry limit is reached, the build is now properly set as a
failure.