Use One Credential Instead of Many
The credential configuration now only supports one credential option,
but this allows authentication against basic auth (username and
password) and DC/OS service accounts (jwt). In the case of the latter,
the credential is updated with the token that was received.

The UI was also simplified, with numerous options now tucked under the
"Advanced" button. All options are still available and take the same
inputs, but the UI is much more compact on initial page load.