Modify the marathon app directly
Prior to this commit, the JSON file on disk was read and then the
resulting JSON object modified according to the configuration in
Jenkins. Finally, that JSON was converted to a Marathon application,
sometimes overwriting fields it should not.

This commit creates the Marathon app from the on-disk JSON file and
then modifies the app according to the Jenkins configuration. This
ensures the app is always correctly formed and leverages more of the
marathon-client library.